Checkstyle and finally putting the max hints back down to 3 for submission. Had to change the whole package name for Glassbox testing to appease the Checkstyle gods.
