+++ src/edu/ipfw/sumfun/Game.java
@@ -420,8 +420,8 @@ public abstract class Game extends Observable {// start Game class
 	}// end setIsActive method
 	
 	public void setGameWon(boolean gameWon) {
-		this.gameWon = gameWon
-;	}
+		this.gameWon = gameWon;
+		}
 	
 	/*
 	 * Acessor method for removeNumAvailable
