+++ src/edu/ipfw/sumfun/Images.java
@@ -7,7 +7,6 @@ import java.awt.Image;
 import java.io.File;
 import java.io.IOException;
 
-
 import javax.imageio.ImageIO;
