+++ src/glassboxtesting/GameBoard.java
@@ -1,4 +1,4 @@
-package Unit;
+package glassboxtesting;
 
 /**
  * The GameBoard class holds and controls the game board
